pyproject.toml
[tool.poetry]
name = "container_service"
version = "0.1.0"
description = ""
authors = ["Jacob Zuliani <jake@burla.dev>"]
packages = [{include = "container_service", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.3"
gunicorn = "^20.1.0"
grpcio = "^1.54.2"
google-cloud-firestore = "^2.11.1"
google-api-python-client = "^2.97.0"
google-cloud-storage = "^2.10.0"
cloudpickle = "^2.2.1"
tblib = "^3.0.0"
google-cloud-logging = "^3.8.0"
pip = "^23.3.1"
install = "^1.3.5"
google-cloud-secret-manager = "^2.16.4"
six = "^1.16.0"
# these are dev dependencies but I want the inatalled in "prod" so I can develop in "prod"
black = "^23.1.0"
pytest = "^7.4.3"
google-cloud-run = "^0.10.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
